+++ tsconfig.json
@@ -31,8 +31,9 @@
     // "baseUrl": "./",                                  /* Specify the base directory to resolve non-relative module names. */
     "paths": {
       /* Specify a set of entries that re-map imports to additional lookup locations. */
-      "@components/*": ["./src/components/*"],
-      "@pluginTypes/*": ["./types/*"]
+      "@/components/*": ["./src/components/*"],
+      "@/helpers/*": ["./src/helpers/*"],
+      "@/pluginTypes/*": ["./types/*"]
     },
     // "rootDirs": [],                                   /* Allow multiple folders to be treated as one when resolving modules. */
     // "typeRoots": [],                                  /* Specify multiple folders that act like './node_modules/@types'. */
